Began some work on the XP system.
Renamed OpenRP_Client.pk3 to OpenRP_Extras.pk3 so it's easier for people to not use the effects/new animations if they don't want to. The .dlls will remain outside of the pk3.

Made qwbitvalues show the admin bitvalues to only the person who used the command, and not the whole server.

git-svn-id: http://openrp.googlecode.com/svn/trunk@15 3c299bc4-26c0-bb6f-2937-d8014291f781