Serverside:
Improved:
RemoveEnt now removes ents based on ID. When you place an effect or spawn an ent, its ID is printed to the console of the effect/ent creator.

All Chat is now designated with <All Chat> Preceding the chat text instead of an @.

The OpenRP client is no longer required to join a server, but of course is strongly recommended to avoid crashes and have the best possible experience.

git-svn-id: http://openrp.googlecode.com/svn/trunk@244 3c299bc4-26c0-bb6f-2937-d8014291f781