Added/modified commands for disguise, amInvisible, and listadmins.
Added /disguise for admins to disguise themselves from the admin list so non-admins can't see them in it.

Renamed amListAdmins to listAdmins, as the command can now be used by non-admins to view admins not in disguise. Admins can see admins who are disguised in the list though.

Readded all amInvisisble functionality.

Made it so you can't emote while force lifting to prevent the endlessly floating up bug. This may just be a temporary workaround, or may be permanent.

git-svn-id: http://openrp.googlecode.com/svn/trunk@296 3c299bc4-26c0-bb6f-2937-d8014291f781